add custom field info to the trello alert
custom fields are a plugin and plugins are not exposed via the
trello api, so requests is used to collect the information

A couple other things cleanups as well related to pep8/pyflakes.
Not all long lines were fixed.

A couple FIXMEs listed as well in case there are multiple plugins
like this in the future.  The current boards in use don't have
multiples at the moment.

Also todo:  add the ability to turn on/off the DFG/RCA info