Add case-insensitive, sub-string, label filtering
2 options for adding label filtering:

At the "channel" level.  This will filter all lists active on
that channel to only cards with a matching labels

At the channel and list level.  This allows setting a filter
that applies only to a specific trello list.

If both are set at the same time, then a card matching either filter
will be displayed

All Clear messages will now be shown for a list when all cards
previously displayed are clean

As part of this fix, an additional benefit is that the All Clear
message will now be shown when DFG filtering is in place.
Previously, this message only showed when a list was completely
clear.  It did not display when the filtered list was cleared.